refactor:重构mooc-authorize部分代码

+++ mooc-authorize2/src/main/java/cn/edu/lingnan/authorize/dao/UserDAO.java
@@ -26,7 +26,7 @@ public class UserDAO extends BaseDAO {
 
         @Override
         public MoocUser mapRow(ResultSet resultSet, int i) throws SQLException {
-            return new MoocUser().setId(resultSet.getInt("id"))
+            return new MoocUser().setId(resultSet.getLong("id"))
                     .setName(resultSet.getString("name"))
                     .setAccount(resultSet.getString("account"))
                     .setPassword(resultSet.getString("password"))
